Move interface to score lib
* Move address setter getter interface to its own interface file
* Add common interfaces as base
