+++ core-contracts/Rebalancing/build.gradle
@@ -25,9 +25,6 @@ repositories {
 dependencies {
     compileOnly Dependencies.javaeeApi
 
-    annotationProcessor Dependencies.javaeeScoreClient
-    compileOnly Dependencies.javaeeScoreClient
-
     implementation Dependencies.javaeeScorex
     implementation project(':score-lib')
 
@@ -46,8 +43,6 @@ optimizedJar {
     duplicatesStrategy = DuplicatesStrategy.EXCLUDE
     from {
         configurations.runtimeClasspath.collect{it.isDirectory() ? it : zipTree(it)}
-    }{
-        exclude "score/*"
     }
 
     enableDebug = false
