+++ score-lib/src/main/java/network/balanced/score/lib/interfaces/addresses/SicxAddress.java
@@ -14,28 +14,16 @@
  * limitations under the License.
  */
 
-import network.balanced.score.dependencies.Dependencies
+package network.balanced.score.lib.interfaces.addresses;
 
-plugins {
-    id 'java'
-}
+import score.Address;
+import score.annotation.External;
 
-version '0.1.0'
+public interface SicxAddress {
 
-repositories {
-    mavenCentral()
-}
+    @External
+    void setSicx(Address _address);
 
-optimizedJar.enabled = false
-
-dependencies {
-    compileOnly Dependencies.javaeeApi
-    compileOnly Dependencies.javaeeScoreClient
-
-    testImplementation Dependencies.junitJupiter
-    testRuntimeOnly Dependencies.junitJupiterEngine
-}
-
-test {
-    useJUnitPlatform()
+    @External(readonly = true)
+    Address getSicx();
 }
